Add test that file source correctly maps to destination
This is to cover the case where files are specified with
src option but no cwd or dest options.

Note that these are really tests of how grunt manages the files
option, but necessary to ensure that the inconsistencies of grunt
processing are captured (e.g. if src is set for an element in
the files array but dest is not, you have to set the dest
explicitly otherwise it is undefined; in other words, the files
objects which grunt produces can have different APIs, dependent
on the content of Gruntfile.js).